Fix macOS build, remove VS 2008-2012 support
Remove references to Boost from docs and scripts
Fix warnings

+++ RenderSystems/Metal/src/OgreMetalWindow.mm
@@ -149,7 +149,7 @@ namespace Ogre
         return (float)mMetalLayer.contentsScale;
 #else
         NSScreen *screen = mMetalView.window.screen ?: [NSScreen mainScreen];
-        return screen.backingScaleFactor;
+        return (float)screen.backingScaleFactor;
 #endif
     }
     //-------------------------------------------------------------------------
